pyproject.toml
[project]
name = "quant-finance-flash-cards"
version = "0.1.0"
description = "Bite-sized revision flash cards for Quantitative Finance exams and interviews, designed for accessibility and clarity."
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "markdown-it-py>=4.0.0",
]

[project.urls]
Repository = "https://github.com/koysor/quant-finance-flash-cards"

[dependency-groups]
dev = [
    "aiohttp>=3.13.3",
]
